move segment key to storage mod
